clean up and merge utils
