+++ python/requirements.txt
@@ -1,5 +1,2 @@
-psutil
-pyyaml
-janus
 werkzeug
 numpy
