+++ python/requirements_test.txt
@@ -2,5 +2,4 @@ pytest==4.4.1
 pytest-cov==2.7.1
 pytest-timeout==1.3.3
 pytest-asyncio==0.10.0
-janus==0.4.0
 numpy==1.15.0
